requirements.txt
flask
requests
newspaper3k
feedparser
playwright
